Remove Helm chart NOTES.txt
Currently after installing the chart the following is printed:

NOTES:
1. Get the application URL by running these commands:
  export POD_NAME=$(kubectl get pods --namespace crossplane-system -l "app.kubernetes.io/name=oam-core-resources,app.kubernetes.io/instance=addon-oam-kubernetes-local" -o jsonpath
="{.items[0].metadata.name}")
  echo "Visit http://127.0.0.1:8080 to use your application"
  kubectl --namespace crossplane-system port-forward $POD_NAME 8080:80

As far as I can tell this is creating a proxy to the OAM controller container,
which I don't think actually serves HTTP (apart from prometheus metrics). I'd
like to remove these notes before we package this addon with Crossplane 0.11 as
they seem a little misleading.

Signed-off-by: Nic Cope <negz@rk0n.org>